Feat:show result video popup
