+++ abnormal-front/package.json
@@ -11,12 +11,14 @@
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.6.1",
     "date-fns": "^2.30.0",
+    "jsmpeg": "^1.0.0",
     "moment": "^2.29.4",
     "react": "^18.2.0",
     "react-datepicker": "^4.24.0",
     "react-dom": "^18.2.0",
     "react-kakao-maps-sdk": "^1.1.24",
     "react-modal": "^3.16.1",
+    "react-player": "^2.14.0",
     "react-router-dom": "^6.18.0",
     "react-scripts": "5.0.1",
     "recoil": "^0.7.7",
