+++ abnormal-front/src/pages/SignupPage.jsx
@@ -1,5 +1,6 @@
 import React, {useState} from 'react';
-import Header from '../components/Header.jsx';
+import { registerLocale, setDefaultLocale } from 'react-datepicker';
+import ko from 'date-fns/locale/ko';
 import PageLayout from '../components/PageLayout.js';
 import styled, { useTheme } from 'styled-components';
 import { useNavigate } from 'react-router-dom';
@@ -9,6 +10,9 @@ import 'react-datepicker/dist/react-datepicker.css';
 import API from "../utils/API.js";
 import { setSelectionRange } from '@testing-library/user-event/dist/utils/index.js';
 import Swal from "sweetalert2";
+
+registerLocale('ko', ko);
+setDefaultLocale('ko');
 
 const SignupForm = styled.form`
   display: flex;
@@ -26,7 +30,6 @@ const SignupForm = styled.form`
   }
 
   
-
   input {
     padding: 8px;
     box-sizing: border-box;
@@ -73,6 +76,12 @@ const SignupForm = styled.form`
     margin-bottom: 20px;
     margin-top: 15px;
     margin-left: 35px;
+  }
+
+  .password-condition {
+    color: red;
+    font-size: 14px;
+    margin-top: 5px;
   }
 `;
 
@@ -123,6 +132,10 @@ const SignUpPage = () => {
   const navigate = useNavigate();
   const [selectedDate, setSelectedDate] = useState(new Date());
   
+  const [passwordCondition, setPasswordCondition] = useState('');
+  const [passwordMatch, setPasswordMatch] = useState(true);
+  const [idAvailabilityMessage, setIdAvailabilityMessage] = useState('');
+
   const [inputValue, setInputValue] = useState({
     userNumber: "",
     employeeNumber: "",
@@ -138,11 +151,50 @@ const SignUpPage = () => {
   const inputChangeHandler = (e, name) => {
     const { value } = e.target;
 
+    // 비밀번호 조건 체크
+    if (name === 'password') {
+      const isPasswordValid = checkPasswordConditions(value);
+      setPasswordCondition(isPasswordValid ? '' : '비밀번호는 영어 소문자, 대문자, 특수문자 1개 이상 포함, 8자 이상이어야 합니다.');
+    }
+
     setInputValue((prevInputValue) => ({
       ...prevInputValue,
       [name]: value,
     }));
+
+    // 비밀번호 확인 체크
+    if (name === 'passwordConfirm') {
+      setPasswordMatch(value === inputValue.password);
+    }
   };
+
+  const checkPasswordConditions = (password) => {
+    const passwordRegex = /^(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&*()_+{}|:"<>?])[A-Za-z\d!@#$%^&*()_+{}|:"<>?]{8,}$/;
+
+    return passwordRegex.test(password);
+  };
+  
+  const checkDuplicateId = async () => {
+    try {
+      const response = await API.post('/users/check-duplicate-id', { userId: inputValue.userId });
+  
+      if (response.data.available) {
+        // 아이디 사용 가능한 경우
+        setIdAvailabilityMessage('사용 가능한 아이디입니다.');
+        // 기타 필요한 로직 추가 가능
+      } else {
+        // 아이디 중복된 경우
+        setIdAvailabilityMessage('이미 존재하는 아이디입니다.');
+        // 기타 필요한 로직 추가 가능
+      }
+    } catch (error) {
+      console.error('서버와의 통신 중 오류 발생', error);
+    }
+  };
+
+  const handleCheckDuplicateId = () => {
+    checkDuplicateId();
+  }
 
   const handleSignUp = async (e) => {
     e.preventDefault()
@@ -209,7 +261,7 @@ const SignUpPage = () => {
                     onChange={(e) => inputChangeHandler(e, 'department')}
                     style={{marginBottom: '8px'}}
                   />
-                    
+                  
                   <label> 아이디</label>
                   <div className='container'>
                     <input 
@@ -220,25 +272,31 @@ const SignUpPage = () => {
                       onChange={(e) => inputChangeHandler(e, 'userId')}
                       style={{width: 600, marginBottom: '18px', marginTop: '15px', marginLeft: '-11.3px' }}
                     />
-                    <button className='doublecheck-button' type='button'>중복확인</button>
+                    <button className='doublecheck-button' type='button' onClick={handleCheckDuplicateId}>중복확인</button>
                   </div>
-                  
+                  { idAvailabilityMessage && <div style={{ color: idAvailabilityMessage.includes('가능') ? 'blue' : 'red' }}>{idAvailabilityMessage}</div> }
+
                   <label>비밀번호</label>
                   <input 
                     type="password" 
                     required 
+                    name="password"
+                    value={inputValue.password}
+                    onChange={(e) => inputChangeHandler(e, 'password')}
                     style={{marginBottom: '8px'}}
                   />
+                  {passwordCondition && <div className="password-condition">{passwordCondition}</div>}
                         
                   <label>비밀번호 확인</label>
                   <input 
                     type="password" 
                     required 
-                    name='password'
-                    value={inputValue.password}
-                    onChange={(e) => inputChangeHandler(e, 'password')}
+                    name='passwordConfirm'
+                    value={inputValue.passwordConfirm}
+                    onChange={(e) => inputChangeHandler(e, 'passwordConfirm')}
                     style={{marginBottom: '8px'}}
                   />
+                  {!passwordMatch && <div className="password-condition">비밀번호가 일치하지 않습니다.</div>}
                         
                   <label>이름</label>
                   <input 
@@ -258,33 +316,27 @@ const SignUpPage = () => {
                     required
                     popperPlacement='right'
                     showPopperArrow={false}
+                    showYearDropdown
+                    showMonthDropdown
                   />
 
                   <label>이메일</label>
-                  <div className='container'>
                     <input 
                       type='email' 
                       required 
                       name='userEmail'
                       value={inputValue.email}
                       onChange={(e) => inputChangeHandler(e, 'userEmail')}
-                      style={{width: 600, marginBottom: '18px', marginTop: '15px', marginLeft: '-11.3px' }}
                     />
-                    <button className='doublecheck-button' type='button'>중복확인</button>
-                  </div>
 
                   <label>휴대전화</label>
-                  <div className='container'>
                     <input 
                       type='text'  
                       required 
-                      name='userPhoneNumber'
+                      name='use	rPhoneNumber'
                       value={inputValue.phoneNumber}
                       onChange={(e) => inputChangeHandler(e, 'userPhoneNumber')}
-                      style={{width: 322, marginBottom: '15px', marginTop: '15px', marginLeft: '-11.3px' }}
                     />
-                    <button className='authentication-button' type='button'>인증</button>
-                  </div>
                   
                   <div style={{marginBottom: '30px'}}></div>
